Bug Fix. MyoAPI sampleWindow contains duplicate entries.
To calculate the mean-absolute-value (MAV) of recent EMG sensor readings, a
window of the most recent samples (i.e., <timestamp, sensor_reading> pairs) is
maintained within the API. However, rather than storing a local copy of the
sensor readings (which are an array of integers), each sample in the
sampleWindow held a reference to its sensor data. The bug was caused because the
same int-array object was re-used for every EMG sensor reading. This resulted in
every sample of the sampleWindow holding a reference to the same array of
sensor readings, and seemingly containing "duplicate" entries. In reality, every
sample held a reference to the most recent array of sensor readings.

Fixed this bug by clone()-ing the array of sensor readings before storing them
into the sampleWindow. Each sample now has a local copy of its associated sensor
readings.