Myo armband can be used to control falldown game.
Falldown can now be controlled with both the keyboard bindings (LEFT, RIGHT,
UP), as well as the Myo armband. The MyoAPI is hidden behind an "EmgManager"
class, which takes care of myo-specific behavior. The Myo gets automatically
calibrated each time the game begins, and motion is controlled as follows:

  - flexion/extension : move left/right (depending on which arm the band is worn)
  - co-flexion : jump

Current sensor readings can be obtained using the EmgManager's poll() method. A
HashMap<String, Float> is returned, mapping the following keys to a value in the
range [0, 1]:

  - "LEFT" : scaling factor for the -x component of motion. Could be controlled
    by either flexion/extension depending on which arm the band is worn.
  - "RIGHT" : scaling factor for the +x component of motion. Could be controlled
    by either flexion/extension depending on which arm the band is worn.
  - "JUMP" : scaling factor for the -y component of motion. Currently takes on
    only discrete values of either 1.0 or 0.0. Controlled by a simultaneous
    co-flexion of both muscles.

The sensor control policy can be adjusted by modifying the
SETTINGS_EMG_CONTROL_POLICY variable within the EmgManager class. Current
options are:

  - "DIFF" : readings are calculated using the difference between flexion and
    extension sensors.
  - "MAX" : only maximum of flexion/extension sensor is considered. other is
    explicitly set to 0.