Bug Fix. Falldown no longer crashes when no myo connected.
Falldown can now be successfully run with our without the myo-armband connected
to your computer. When connected, the ball can be controlled by either the myo
or the keyboard. When disconnected, the ball is only controlled by the keyboard.

Rather than directly instantiating an EmgManager object, a factory method is
used instead which replaces the EmgManager with a "Null Object" equivalent with
stubbed out methods when no myo armband is detected.